requirements/local.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements/local.txt requirements/local.in
#
alabaster==0.7.12
    # via sphinx
appnope==0.1.2
    # via ipython
argon2-cffi==21.3.0
    # via -r requirements/./base.txt
argon2-cffi-bindings==21.2.0
    # via
    #   -r requirements/./base.txt
    #   argon2-cffi
asgiref==3.5.0
    # via
    #   -r requirements/./base.txt
    #   django
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via pytest
babel==2.9.1
    # via sphinx
backcall==0.2.0
    # via ipython
black==21.12b0
    # via
    #   -r requirements/local.in
    #   ipython
boto3==1.20.42
    # via
    #   -r requirements/./base.txt
    #   ssm-parameter-store
botocore==1.23.42
    # via
    #   -r requirements/./base.txt
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via
    #   -r requirements/./base.txt
    #   whitenoise
certifi==2021.10.8
    # via
    #   -r requirements/./base.txt
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via
    #   -r requirements/./base.txt
    #   argon2-cffi-bindings
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.10
    # via
    #   -r requirements/./base.txt
    #   requests
click==8.0.3
    # via
    #   -r requirements/./base.txt
    #   black
    #   pip-tools
    #   rq
colorama==0.4.4
    # via pytest-watch
coloredlogs==15.0.1
    # via -r requirements/local.in
coverage[toml]==6.2
    # via
    #   -r requirements/local.in
    #   pytest-cov
    #   pytest-testmon
croniter==1.2.0
    # via
    #   -r requirements/./base.txt
    #   rq-scheduler
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
deprecated==1.2.13
    # via
    #   -r requirements/./base.txt
    #   redis
distlib==0.3.4
    # via virtualenv
django==4.0.1
    # via
    #   -r requirements/./base.txt
    #   django-anymail
    #   django-debug-toolbar
    #   django-extensions
    #   django-rq
django-anymail[mailgun]==8.5
    # via -r requirements/./base.txt
django-debug-toolbar==3.2.2
    # via -r requirements/local.in
django-environ==0.8.1
    # via -r requirements/./base.txt
django-extensions==3.1.5
    # via -r requirements/local.in
django-rq==2.5.1
    # via -r requirements/./base.txt
django-test-migrations==1.2.0
    # via -r requirements/local.in
django-waffle==2.3.0
    # via -r requirements/./base.txt
docopt==0.6.2
    # via pytest-watch
docutils==0.17.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
executing==0.8.2
    # via stack-data
factory-boy==3.2.1
    # via -r requirements/local.in
faker==11.3.0
    # via factory-boy
fancycompleter==0.9.1
    # via pdbpp
filelock==3.4.2
    # via virtualenv
flake8==4.0.1
    # via
    #   -r requirements/local.in
    #   flake8-docstrings
flake8-docstrings==1.6.0
    # via -r requirements/local.in
gunicorn==20.1.0
    # via -r requirements/./base.txt
humanfriendly==10.0
    # via coloredlogs
identify==2.4.5
    # via pre-commit
idna==3.3
    # via
    #   -r requirements/./base.txt
    #   requests
imagesize==1.3.0
    # via sphinx
importlib-metadata==4.10.1
    # via sphinx
iniconfig==1.1.1
    # via pytest
ipdb==0.13.9
    # via -r requirements/local.in
ipython==8.0.1
    # via ipdb
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via sphinx
jmespath==0.10.0
    # via
    #   -r requirements/./base.txt
    #   boto3
    #   botocore
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
packaging==21.3
    # via
    #   -r requirements/./base.txt
    #   pytest
    #   pytest-sugar
    #   redis
    #   sphinx
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pdbpp==0.10.3
    # via -r requirements/local.in
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.4.0
    # via -r requirements/local.in
platformdirs==2.4.1
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.17.0
    # via -r requirements/local.in
prompt-toolkit==3.0.24
    # via ipython
psycopg2-binary==2.9.3
    # via -r requirements/./base.txt
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via
    #   -r requirements/./base.txt
    #   cffi
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.4.0
    # via flake8
pygments==2.11.2
    # via
    #   ipython
    #   pdbpp
    #   sphinx
pyparsing==3.0.7
    # via
    #   -r requirements/./base.txt
    #   packaging
pyrepl==0.9.0
    # via fancycompleter
pytest==6.2.5
    # via
    #   -r requirements/local.in
    #   pytest-cov
    #   pytest-django
    #   pytest-pspec
    #   pytest-sugar
    #   pytest-testmon
    #   pytest-watch
pytest-cov==3.0.0
    # via -r requirements/local.in
pytest-django==4.5.2
    # via -r requirements/local.in
pytest-pspec==0.0.4
    # via -r requirements/local.in
pytest-sugar==0.9.4
    # via -r requirements/local.in
pytest-testmon==1.2.2
    # via -r requirements/local.in
pytest-watch==4.2.0
    # via -r requirements/local.in
python-dateutil==2.8.2
    # via
    #   -r requirements/./base.txt
    #   botocore
    #   croniter
    #   faker
    #   rq-scheduler
    #   time-machine
pytz==2021.3
    # via babel
pyyaml==6.0
    # via pre-commit
redis==4.1.1
    # via
    #   -r requirements/./base.txt
    #   django-rq
    #   rq
requests==2.27.1
    # via
    #   -r requirements/./base.txt
    #   django-anymail
    #   requests-mock
    #   sphinx
requests-mock==1.9.3
    # via -r requirements/local.in
rq==1.10.1
    # via
    #   -r requirements/./base.txt
    #   django-rq
    #   rq-scheduler
rq-scheduler==0.11.0
    # via -r requirements/./base.txt
s3transfer==0.5.0
    # via
    #   -r requirements/./base.txt
    #   boto3
sentry-sdk==1.5.3
    # via -r requirements/./base.txt
six==1.16.0
    # via
    #   -r requirements/./base.txt
    #   pytest-pspec
    #   python-dateutil
    #   requests-mock
    #   virtualenv
snowballstemmer==2.2.0
    # via
    #   pydocstyle
    #   sphinx
sphinx==4.4.0
    # via
    #   -r requirements/./docs.in
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.0.0
    # via -r requirements/./docs.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.2
    # via
    #   -r requirements/./base.txt
    #   django
    #   django-debug-toolbar
ssm-parameter-store==19.11.0
    # via -r requirements/./base.txt
stack-data==0.1.4
    # via ipython
termcolor==1.1.0
    # via pytest-sugar
text-unidecode==1.3
    # via faker
time-machine==2.6.0
    # via -r requirements/local.in
toml==0.10.2
    # via
    #   ipdb
    #   pre-commit
    #   pytest
tomli==1.2.3
    # via
    #   black
    #   coverage
    #   pep517
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.0.1
    # via
    #   black
    #   django-test-migrations
urllib3==1.26.8
    # via
    #   -r requirements/./base.txt
    #   botocore
    #   requests
    #   sentry-sdk
virtualenv==20.13.0
    # via pre-commit
watchdog==2.1.6
    # via pytest-watch
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.2
    # via -r requirements/local.in
wheel==0.37.1
    # via pip-tools
whitenoise[brotli]==5.3.0
    # via -r requirements/./base.txt
wmctrl==0.4
    # via pdbpp
wrapt==1.13.3
    # via
    #   -r requirements/./base.txt
    #   deprecated
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
